+++ gazebo/gui/building/BuildingEditorPalette.cc
@@ -48,6 +48,7 @@ BuildingEditorPalette::BuildingEditorPalette(QWidget *_parent)
 
   // Brushes (button group)
   brushes = new QButtonGroup();
+  connect(brushes, SIGNAL(buttonClicked(int)), this, SLOT(OnBrush(int)));
 
   QSize toolButtonSize(100, 100);
   QSize iconSize(65, 65);
@@ -135,7 +136,35 @@ BuildingEditorPalette::BuildingEditorPalette(QWidget *_parent)
     brushes->addButton(colorButton, i);
     colorsLayout->addWidget(colorButton, 0, i);
   }
-  connect(brushes, SIGNAL(buttonClicked(int)), this, SLOT(OnColor(int)));
+
+  // Textures
+  QLabel *texturesLabel = new QLabel(tr(
+       "<font size=4 color='white'>Add Texture</font>"));
+
+  QGridLayout *texturesLayout = new QGridLayout;
+  std::vector<QString> textureButtonTextList;
+
+  this->textureList.push_back(":wood.jpg");
+  textureButtonTextList.push_back("Wood");
+  this->textureList.push_back(":tiles.jpg");
+  textureButtonTextList.push_back("Tiles");
+  this->textureList.push_back(":bricks.png");
+  textureButtonTextList.push_back("Bricks");
+
+  for (unsigned int i = 0; i < this->textureList.size(); ++i)
+  {
+    QToolButton *textureButton = new QToolButton(this);
+    textureButton->setFixedSize(70, 70);
+    textureButton->setCheckable(true);
+    textureButton->setChecked(false);
+    textureButton->setToolButtonStyle(Qt::ToolButtonTextUnderIcon);
+    textureButton->setIcon(QPixmap(this->textureList[i]).scaled(
+        QSize(90, 90), Qt::IgnoreAspectRatio));
+    textureButton->setText(textureButtonTextList[i]);
+    textureButton->setIconSize(QSize(40, 40));
+    brushes->addButton(textureButton, brushes->buttons().size());
+    texturesLayout->addWidget(textureButton, 0, i);
+  }
 
   // Import button
   QPushButton *importImageButton = new QPushButton(tr("Import"),
@@ -161,6 +190,8 @@ BuildingEditorPalette::BuildingEditorPalette(QWidget *_parent)
   mainLayout->addLayout(featuresLayout);
   mainLayout->addWidget(colorsLabel);
   mainLayout->addLayout(colorsLayout);
+  mainLayout->addWidget(texturesLabel);
+  mainLayout->addLayout(texturesLayout);
   mainLayout->addItem(new QSpacerItem(10, 20, QSizePolicy::Expanding,
                       QSizePolicy::Minimum));
   mainLayout->addLayout(buttonsLayout);
@@ -181,7 +212,7 @@ BuildingEditorPalette::BuildingEditorPalette(QWidget *_parent)
       gui::editor::Events::ConnectCreateBuildingEditorItem(
       boost::bind(&BuildingEditorPalette::OnCreateEditorItem, this, _1)));
 
-  // All buttons must be added after the color buttons
+  // All buttons must be added after the color and texture buttons
   brushes->addButton(wallButton, brushes->buttons().size());
   brushes->addButton(windowButton, brushes->buttons().size());
   brushes->addButton(doorButton, brushes->buttons().size());
@@ -268,6 +299,7 @@ void BuildingEditorPalette::OnNameChanged(const QString &_name)
 void BuildingEditorPalette::OnCreateEditorItem(const std::string &_mode)
 {
   gui::editor::Events::colorSelected(QColor::Invalid);
+  gui::editor::Events::textureSelected(QString(""));
 
   if (_mode.empty() || this->currentMode == _mode)
   {
@@ -285,15 +317,27 @@ void BuildingEditorPalette::OnCreateEditorItem(const std::string &_mode)
 }
 
 /////////////////////////////////////////////////
+void BuildingEditorPalette::OnBrush(int _buttonId)
+{
+  if (_buttonId < static_cast<int>(colorList.size()))
+  {
+    this->OnColor(_buttonId);
+  }
+  else if (_buttonId < static_cast<int>(colorList.size()) +
+                       static_cast<int>(textureList.size()))
+  {
+    this->OnTexture(_buttonId - static_cast<int>(colorList.size()));
+  }
+  else
+  {
+    gzwarn << "Brushes other than color and texture are handled elsewhere."
+           << std::endl;
+  }
+}
+
+/////////////////////////////////////////////////
 void BuildingEditorPalette::OnColor(int _buttonId)
 {
-  // A button which is not color
-  if (_buttonId >= static_cast<int>(colorList.size()))
-  {
-    gzwarn << "Brushes other than color are handled elsewhere." << std::endl;
-    return;
-  }
-
   std::ostringstream colorStr;
   colorStr << "color_" << _buttonId;
   QColor color = this->colorList[_buttonId];
@@ -305,6 +349,28 @@ void BuildingEditorPalette::OnColor(int _buttonId)
     QPixmap colorCursor(30, 30);
     colorCursor.fill(color);
     QApplication::setOverrideCursor(QCursor(colorCursor));
+  }
+  else
+  {
+    gui::editor::Events::createBuildingEditorItem(std::string());
+  }
+}
+
+/////////////////////////////////////////////////
+void BuildingEditorPalette::OnTexture(int _buttonId)
+{
+  std::ostringstream textureStr;
+  textureStr << "texture_" << _buttonId;
+  QString texture = this->textureList[_buttonId];
+  if (this->currentMode != textureStr.str())
+  {
+    gui::editor::Events::textureSelected(texture);
+    this->currentMode = textureStr.str();
+
+    QPixmap textureCursor(this->textureList[_buttonId]);
+    textureCursor = textureCursor.scaled(QSize(30, 30), Qt::IgnoreAspectRatio,
+        Qt::SmoothTransformation);
+    QApplication::setOverrideCursor(textureCursor);
   }
   else
   {